Combining withSaml and withAuthentication
