+++ src/components/withAuthentication/tests/withAuthentication.test.jsx
@@ -3,6 +3,7 @@ import { mount } from 'enzyme';
 import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
 import configureMockStore from 'redux-mock-store';
+import { StaticQuery } from 'gatsby';
 
 import apiClient from '../../../data/apiClient';
 
@@ -20,12 +21,24 @@ apiClient.ensurePublicOrAuthenticationAndCookies = (_, callback) => {
 };
 
 describe('<withAuthentication />', () => {
+  beforeEach(() => {
+    StaticQuery.mockImplementationOnce(({ render }) => (
+      render({
+        site: {
+          siteMetadata: {
+            providerSlug: 'test-saml',
+          },
+        },
+      })
+    ));
+  });
+
   it('does not render anything on initial paint', () => {
     const MyComponent = () => <div />;
     const AuthenticatedComponent = withAuthentication(MyComponent);
     const wrapper = mount((
       <Provider store={store}>
-        <AuthenticatedComponent location={{ pathname: '/' }} providerSlug="test-saml" />
+        <AuthenticatedComponent location={{ pathname: '/' }} />
       </Provider>
     ));
     expect(wrapper.html()).toBeNull();
