+++ src/pages/index.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import { IntlProvider } from 'react-intl';
 
 import withAuthentication from '../components/withAuthentication';
-import withSaml from '../components/withSaml';
 import Layout from '../components/Layout/Layout';
 import DashboardHome from '../components/DashboardHome/DashboardHome';
 
@@ -14,4 +13,4 @@ const IndexPage = () => (
   </IntlProvider>
 );
 
-export default withSaml(withAuthentication(IndexPage));
+export default withAuthentication(IndexPage);
